Fixed compile errors from warning flag updates. fixed race in cleanup for socket handler, cleaned up darwin sssl handler.

+++ source/bsd/kqueue_event_loop.c
@@ -759,7 +759,7 @@ static void s_event_thread_main(void *user_data) {
 
                 /* Drain whatever data was written to the signaling pipe */
                 uint32_t read_whatever;
-                while (read(kevent->ident, &read_whatever, sizeof(read_whatever)) > 0) {
+                while (read((int)kevent->ident, &read_whatever, sizeof(read_whatever)) > 0) {
                 }
 
                 continue;
